Move real api integration tests to separate file. Remove integration tests from coverage.
