+++ docker-compose.override.yml
@@ -11,6 +11,15 @@
 
 version: "3.8"
 services:
+  fakeapi:
+    build:
+      dockerfile: ./docker/json_server/Dockerfile
+    command: -H 0.0.0.0 -p 3000 -w db.json
+    ports:
+      - "3000:3000"
+    networks:
+      - webnet
+
   web:
     ports:
       # We only bind ports directly in development:
