+++ setup.cfg
@@ -118,6 +118,7 @@ plugins =
 omit =
   # Is not reported, because is imported during setup:
   server/settings/components/logging.py
+  tests/test_apps/test_identity/test_integration_with_api.py
   tests/conftest.py
 
 [mypy]
