+++ tests/test_apps/test_identity/test_placeholder.py
@@ -1,4 +1,3 @@
-import os
 from typing import TYPE_CHECKING, Generator
 
 import pytest
@@ -8,17 +7,12 @@ from server.apps.identity.infrastructure.services.placeholder import (
     UserResponse,
 )
 from server.apps.identity.models import User
-from server.settings.components.placeholder import PLACEHOLDER_API_URL
 
 if TYPE_CHECKING:
     from plugins.identity.external_api import ExternalApiUserResponse
 
 
-pytestmark = [
-    pytest.mark.django_db(),
-]
-
-
+@pytest.mark.django_db()
 def test_lead_create_using_mocked_api(
     user_with_birthdate: User,
     external_api_mock: Generator['ExternalApiUserResponse', None, None],
@@ -30,26 +24,3 @@ def test_lead_create_using_mocked_api(
     assert response == UserResponse(
         **external_api_mock,  # type: ignore[arg-type]
     )
-
-
-@pytest.mark.slow()
-@pytest.mark.timeout(10)
-@pytest.mark.parametrize(
-    'url',
-    [
-        PLACEHOLDER_API_URL,
-        pytest.param(
-            'http://fakeapi:3000',
-            marks=pytest.mark.skipif(
-                os.getenv('DJANGO_ENV') == 'production',
-                reason='This test is not supposed to be run in production',
-            ),
-        ),
-    ],
-)
-def test_lead_create_using_real_api(
-    user_with_birthdate: User,
-    url: str,
-) -> None:
-    """Tests LeadCreate for user on real API and local json-server."""
-    LeadCreate(url, 10)(user=user_with_birthdate)
